requirements.txt
python-Binance
TA-Lib
numpy
websocket_client